+++ src/main/java/mediathek/tool/swing/ThreadCheckingRepaintManager.kt
@@ -3,7 +3,7 @@
  *
  * Copyright 2002 - 2006 JIDE Software Inc. All rights reserved.
  */
-package mediathek.tool
+package mediathek.tool.swing
 
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
